+++ ios/Todo-Assignment/Sources/Models/TaskModel.swift
@@ -63,7 +63,7 @@ class TaskModel: RealmModel<Task> {
     /// - parameter title: タイトル(タスク名)
     /// - parameter withID: エンティティに与えるID(省略時は自動的に採番)
     /// - returns: 新しいエンティティ
-    func create(title: String, withID id: Int64? = nil) -> Entity {
+    func create(title: String, withID id: Int? = nil) -> Entity {
         let ret = self.create(withID: id)
         ret.title = title
         ret.date  = Date.today()
